chore: savepoint

+++ libtransmission/transmission.h
@@ -587,8 +587,8 @@ bool tr_sessionGetDeleteSource(tr_session const*);
 tr_priority_t tr_torrentGetPriority(tr_torrent const*);
 void tr_torrentSetPriority(tr_torrent*, tr_priority_t);
 
-void tr_sessionSetAntiBruteForceThreshold(tr_session*, int bad_requests);
-int tr_sessionGetAntiBruteForceThreshold(tr_session const*);
+void tr_sessionSetAntiBruteForceThreshold(tr_session*, size_t bad_requests);
+size_t tr_sessionGetAntiBruteForceThreshold(tr_session const*);
 
 void tr_sessionSetAntiBruteForceEnabled(tr_session*, bool enabled);
 bool tr_sessionGetAntiBruteForceEnabled(tr_session const*);
